organized the folder structure of components
