+++ src/components/Common/QuotedImage.tsx
