+++ src/components/CulturePage/AvailablePositions.tsx
@@ -3,9 +3,9 @@ import type { FC } from 'react';
 
 import { atMinXL, spacing } from '~/theme';
 
-import { Heading } from '../Heading';
-import { Link } from '../Link';
-import { Text } from '../Text';
+import { Heading } from '../Common/Heading';
+import { Link } from '../Common/Link';
+import { Text } from '../Common/Text';
 
 export const AvailablePositions: FC = (props) => {
   return (
