+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export * from './ApproachPage';
 export * from './BatteryXTPage';
 export * from './CareersPage';
-export * from './Card';
+export * from './Common/Card';
 export * from './CaseStudy/CaseStudyBottomNav';
 export * from './CaseStudy/CaseStudyHeaderPane';
 export * from './CaseStudy/CaseStudyHero';
@@ -9,27 +9,27 @@ export * from './CaseStudy/CaseStudyIntroContentPane';
 export * from './CaseStudy/CaseStudyResultsPane';
 export * from './CaseStudy/CaseStudySummaryPane';
 export * from './ColorSnapVisualizerPage';
-export * from './Container';
+export * from './Common/Container';
 export * from './ContactPage';
 export * from './CulturePage';
 export * from './Custom404Page';
 export * from './Custom500Page';
-export * from './DynamicImage';
-export * from './GradientCircle';
-export * from './Header';
-export * from './Heading';
+export * from './Unused/DynamicImage';
+export * from './Common/GradientCircle';
+export * from './Layouts/Header';
+export * from './Common/Heading';
 export * from './HomePage';
 export * from './icons';
 export * from './ColorSnapVisualizerPage/ColorSnapImagePanel';
-export * from './Layout';
-export * from './Link';
-export * from './MobileMenu';
+export * from './Layouts/Layout';
+export * from './Common/Link';
+export * from './Layouts/MobileMenu';
 export * from './NetJetsPage';
-export * from './OpenGraphImage';
-export * from './PageHero';
-export * from './QuotedImage';
+export * from './Unused/OpenGraphImage';
+export * from './Common/PageHero';
+export * from './Common/QuotedImage';
 export * from './ScopeBuilderPage';
-export * from './Text';
-export * from './useAnimationWhileVisible';
-export * from './useMouseAnimationWhileVisible';
+export * from './Common/Text';
+export * from './Hooks/useAnimationWhileVisible';
+export * from './Hooks/useMouseAnimationWhileVisible';
 export * from './WorkPage';
